:card_file_box:feat: id Long으로 변경 #2

+++ backend/focus/src/main/java/com/bb/focus/db/entity/Evaluator.java
@@ -19,7 +19,8 @@ public class Evaluator {
     @Column(name = "evaluator_id")
     private int id;
 
-    private int companyAdminId;
+    private Long companyAdminId;
+
     private String userId;
     private String pwd;
     private String name;
